Create a function existence registry
With the signature module, it is now a straight forward task
to generate a registry of all known function signatures.
We check that these signatures all exist.

One file contains all signatures for one platform and version.

The test is only activated when run in the CI system.
An initial call creates the expected file as output and raises
an error. The result can then be picked up from the error log
and added to the repository.

Done: linux2 5.6.4
Done: darwin 5.6.4
Done: win32 5.6.4

Task-number: PYSIDE-510
Change-Id: I4f406cf72d25fdd2336814f6f20129079b8be54f
Reviewed-by: Christian Tismer <tismer@stackless.com>