Fix signature registry on Python2 with .pyc files
There was a problem on Python2 when ci was activated.
Because there are .pyc files in the same folder,
a leftover .pyc file would be imported and lead to weird results.

This problem is not recognized now, but would have effects
when we turn on the multiple testing. The intended behavior
is that a tests generates an error and a listing once and
succeeds for the repeated test runs. This worked in Python3.

Now this works the same with Python2.

Task-number: PYSIDE-510
Change-Id: Id892715faa8eee1322b28c7e109f3b0b7329f12c
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>