Fix cmake scope for locally installed modules
If modules are installed on the system but not in
the local Qt installation, cmake will find them
and try to build, but this will lead to errors.

Comparing the QT_INCLUDE_DIR and the path where
the package was found, we can filter the modules
that were found on a different directory.

Change-Id: I463172b7081c191fb303387f98a495dd675e0486
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>