Add code to invalidate objs on QListWidget.clear
As reported on PYSIDE-264, when an item of an already cleared
QListWidget was being accessed, a segfault happened when trying
to access this deleted data.
Due to the lack of an invalidation process of the python objects
when clear() was called, the generated validation step before accessing
the data had no effect.
This was solved injecting code to set their parents to NULL, and
invalidating them.
The outcome of trying to access deleted data then will be a RuntimeError,
instead of a segfault.
A test case is provided.

Task-number: PYSIDE-264
Change-Id: If52dd85827500c96a078a8f9d61921a275fb28f9
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>