Registry: Trigger re-generation of files for Qt3D
This is phase 2.
The files are generated. I only changed the date in the license section.
After this check-in, the update is done.

From Phase 1:

The function registry does not contain the new Qt3D module.
This produces no error because the test is configured to only
break on missing functions but not on new unknown functions.

We provoke a reaction of the system by removing of the 5.9
registry files. The system will generate an error once and produce
the desired output.
But because of the multiple testing, the test will succeed because
the generated file exist on the second run and therefore the test
will succeed as a flaky test.
There is only one run necessary for all platforms at once.
A second check-in will then do the update with the generated data.

I also had to fix the testrunner to produce a good listing without labels.

Task-number: PYSIDE-487
Change-Id: I06a73d244ce306977fd16223eec4dc491fff3429
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>