Check slot-signal association when connecting
When connecting a signal with a slot there is a process
to associate the proper signal signature, but the slot
signature was not verified.
This missing verification step lead to wrongly associate
the slots and the signal signatures, for example:

def on_clicked(checked=True):
    ...
QGroupBox.clicked.connect(on_clicked)

will wrongly connect the slot "on_clicked" with the
signal "clicked()" (without any argument),
when the proper signal is "clicked(bool)".
This can be solved by manually specifying the arguments:

QGroupBox.clicked[bool].connect(self.clicked)

We can add an additional verification step
to associate the proper signal if the slot has
a certain number of arguments.

There is an existing test that checks the compatibility
of this change with all the ways to connect
signals and slots.
A few additional cases were added.

Task-number: PYSIDE-104
Change-Id: Ic5b06fa3bb91903f7d506e0e2c52a6f7d3dc4570
Reviewed-by: Christian Tismer <tismer@stackless.com>