Add condition to qmimedatabase_test
An assert related to a Mimetype comment fails if the language
is unset or is not English.
Since this assert is not critical to the functionality of the module,
we can remove it or add a simple condition to test it only when
the system's language is English.
The last one was applied.

Task-number: PYSIDE-575
Change-Id: Ida85268e25522e406878dd6ac4e5e70852f8d8e4
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Christian Tismer <tismer@stackless.com>