Signature: make the parser more complete for 5.11
The parser regex could not handle angle bracket pairs with commas in it.

This is needed for template parameter lists. When they contain commata
between the angle brackets, the parser did not recognize that.

This fix allows for one level of angle brackets with whatever content.
It will probably be needed in 5.11, but the syntax that the regex recognizes
should always be complete.

I had a hard time to understand this split regex again, so I added some
more documentation, and it should now be simple to extend it even more.

Task-number: PYSIDE-510
Task-number: PYSIDE-616
Change-Id: Ic854852f35af8b4526a63ffe920f2c01204c1f31
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>