Improve qt.conf registration handling
When PyInstaller is used to deploy a PySide2 application, it changes
the layout of the copied PySide2 files (no Qt subfolder, all shared
libraries are copied next to tehe generated executable). In that case
using the internal qt.conf won't work.

Detect if there a exists a qt.conf file next to the executable, and
use that file instead.

Note that this won't work when the executable path has unicode
characters in conjunction with Python 2 and Windows, but that is an
unsupported config anyway (due to the mixing MSVC issue).

Also add a logging category to ease figuring out which qt.conf file
is used if a need ever arises to do so.

Task-number: PYSIDE-642
Change-Id: I1260cbc13e5e62be72c4ed9c64c2aa5905d2e9c6
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>