Use PyLong instead of PyInt for QVariant
Python 3 uses PyLong for integer manipulation,
and since it is a valid type in Python 2 we can
use it instead to have better compatibility,
and avoid integer overflows.

We keep the PyInt entry to avoid breaking
other sections of the code when this type is required.

Added a couple of PyInt_* alias to use PyLong_* ones
instead.

Task-number: PYSIDE-648
Change-Id: Id6af66050c5ab81fc51c1c07b1c9422a954b81c0
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Christian Tismer <tismer@stackless.com>