Fix qtattributionsscannertorst.py
The tool was originally written to handle a license embedded
in a Python file. This has changed, the license now exists as
a text file.

Rewrite the tool to concatenate the output piece-wise instead
of using fixed formats to be able to cope with missing URLs and
versions. Add the copyrights, which were previously missing.

Task-number: PSYIDE-363
Change-Id: I5ec0bd7a4b81e9f794f56f6b2b09062c331936ae
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>