Allow building just QtCore bindings
This change introduces tracking of modules for which bindings will not
be built. This allows us to add conditions on disabled modules,
and thus disable certain tests.

Thus we disable pysidetest test when its dependencies are not met.

Due to this, we can now builtd bindings only for QtCore,
allowing faster development iteration when touching only QtCore.

Note that this only affects which module bindings will be created.

Shiboken itself still requires QtCore, QtXml and QtXmlPatterns,
as does pyside2-rcc.

Change-Id: I75084a1741e7f4c3594e43af0fd9668a0e969c56
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>