Implement Proper Name Mangling
When types have attributes starting with two underscores but
ending with at most one, Python uses name mangling to create
a unique private variable.

PySide needs to obey this rule in the tp_getattro methods.
We implemented it as an optimized _Pep_PrivateMangle function that
solves the problem internally without exposing the _Py_Mangle
function.

Remark: I think the exclusion of the _Py_Mangle function is another
oversight in the Limited API.

Task-number: PYSIDE-772
Change-Id: I0bfc2418dae439e963a16e37443f2099c6980696
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>