Reset the Signature Registry after the Total Scan
After scanning the whole COIN test website, a lot of changes were
found which we missed by manually inspectation.

This update contains the essence of all changes which accumulated
on the server. They were sorted by date, and the latest version
was taken. Then, the license headers were updated, because
that had been done already on the repository.

Update
------

This check-in did not work on 2018-09-07.
The reason was that we changed the PySide version to 5.11.2,
which was not yet in my update.

While re-doing the generation, I realized that we will not
automatically get the 5.11.2 version, because of the "fallback".
feature.

To reset the existence file, we need to remove them once.
After the checkin, all the files will be freshly created.

I propose also to disable the "fallback" feature. In the
presence of the multiple COIN testing, it is no more
necessary.

Change-Id: I43af314f7bd233a4190762c7204e8aecf81bfe72
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>