Simplify and correct the new_inherited_functions test
The test was written when PySide2 did not have the
"from PySide2 import *" feature. It was written with "try..except"
constructs to find the translated modules.

This did not work correctly due to the path settings in ctest.
When not all modules are selected, they are still partially importable!

For instance, PySide2.QtPrintSupport could be imported, although
it was not compiled, and then complained since a function was not found.

This is possible, because python 3 supports module directories without
an __init__.py file. (I was always against that "feature".)

Instead of adjusting the path setting, this patch uses the above feature
and does not rely on "importability".

Task-number: PYSIDE-331
Change-Id: I9aed4971f10f423918525cf411fd7ff713f40e48
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>